requirements.txt
fastapi==0.115.6
pydantic==2.10.3
openai==1.57.2
chromadb==0.5.23
pdf2image==1.17.0
pytesseract==0.3.13
pillow==11.0.0
python-dotenv==1.0.0
uvicorn==0.32.1
langchain-openai==0.2.12
python-multipart==0.0.19
langchain==0.3.11
langchain-chroma==0.1.4
easyocr==1.7.2